refactor(indexeddb): rename serializer types modules to indexed_types
