+++ crates/matrix-sdk-indexeddb/src/event_cache_store/serializer/indexed_types.rs
