+++ crates/matrix-sdk-indexeddb/src/media_store/serializer/constants.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License
 
-use crate::media_store::serializer::types::IndexedMediaContentSize;
+use crate::media_store::serializer::indexed_types::IndexedMediaContentSize;
 
 /// An [`IndexedMediaContentSize`] set to it's minimal value - i.e., `0`.
 ///
