+++ crates/matrix-sdk-indexeddb/src/media_store/serializer/indexed_types.rs
